+++ ts/components/ui/IOScrollView.tsx
@@ -44,6 +44,7 @@ import Animated, {
 } from "react-native-reanimated";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 import { WithTestID } from "../../types/WithTestID";
+import { useStatusAlertProps } from "../../hooks/useStatusAlertProps";
 
 export type IOScrollViewActions =
   | {
@@ -160,6 +161,7 @@ export const IOScrollView = ({
   contentContainerStyle,
   testID
 }: IOScrollView) => {
+  const alertProps = useStatusAlertProps();
   const theme = useIOTheme();
 
   /* Navigation */
@@ -243,6 +245,13 @@ export const IOScrollView = ({
     )
   }));
 
+  const ignoreSafeAreaMargin = React.useMemo(() => {
+    if (alertProps !== undefined) {
+      return true;
+    }
+    return headerConfig?.ignoreSafeAreaMargin;
+  }, [headerConfig?.ignoreSafeAreaMargin, alertProps]);
+
   /* Set custom header with `react-navigation` library using
      `useLayoutEffect` hook */
 
@@ -255,12 +264,22 @@ export const IOScrollView = ({
     if (headerConfig) {
       navigation.setOptions({
         header: () => (
-          <HeaderSecondLevel {...headerConfig} scrollValues={scrollValues} />
+          <HeaderSecondLevel
+            {...headerConfig}
+            ignoreSafeAreaMargin={ignoreSafeAreaMargin}
+            scrollValues={scrollValues}
+          />
         ),
         headerTransparent: headerConfig.transparent
       });
     }
-  }, [headerConfig, navigation, scrollPositionAbsolute, snapOffset]);
+  }, [
+    headerConfig,
+    navigation,
+    scrollPositionAbsolute,
+    snapOffset,
+    ignoreSafeAreaMargin
+  ]);
 
   const RefreshControlComponent = refreshControlProps ? (
     <RefreshControl {...refreshControlProps} />