+++ ts/features/itwallet/common/store/actions/preferences.ts
@@ -4,4 +4,10 @@ export const itwCloseFeedbackBanner = createStandardAction(
   "ITW_CLOSE_FEEDBACK_BANNER"
 )();
 
-export type ItwPreferencesActions = ActionType<typeof itwCloseFeedbackBanner>;
+export const itwCloseDiscoveryBanner = createStandardAction(
+  "ITW_CLOSE_DISCOVERY_BANNER"
+)();
+
+export type ItwPreferencesActions =
+  | ActionType<typeof itwCloseFeedbackBanner>
+  | ActionType<typeof itwCloseDiscoveryBanner>;