+++ ts/features/payments/checkout/store/reducers/index.ts
@@ -16,7 +16,6 @@ import { Action } from "../../../../../store/actions/types";
 import { NetworkError } from "../../../../../utils/errors";
 import { getSortedPspList } from "../../../common/utils";
 import { WalletPaymentStepEnum } from "../../types";
-import { FaultCodeCategoryEnum } from "../../types/PaymentGenericErrorAfterUserCancellationProblemJson";
 import { FaultCodeCategoryEnum as PaymentMethodNotAvailableEnum } from "../../types/PspPaymentMethodNotAvailableProblemJson";
 import { WalletPaymentFailure } from "../../types/WalletPaymentFailure";
 import {
@@ -265,20 +264,6 @@ const reducer = (
       return {
         ...state,
         transaction: pot.toError(state.transaction, action.payload)
-      };
-    case getType(paymentsDeleteTransactionAction.failure):
-      return {
-        ...state,
-        transaction: pot.toError(
-          state.transaction,
-          action.payload.kind === "generic"
-            ? {
-                faultCodeCategory:
-                  FaultCodeCategoryEnum.PAYMENT_GENERIC_ERROR_AFTER_USER_CANCELLATION,
-                faultCodeDetail: ""
-              }
-            : action.payload
-        )
       };
 
     // Authorization url