+++ ts/features/payments/details/saga/handleDeleteWalletDetails.ts
@@ -1,22 +1,15 @@
-import { put } from "typed-redux-saga/macro";
 import * as E from "fp-ts/lib/Either";
+import { put } from "typed-redux-saga/macro";
 import { ActionType } from "typesafe-actions";
+import { getGenericError, getNetworkError } from "../../../../utils/errors";
+import { readablePrivacyReport } from "../../../../utils/reporters";
+import { WalletClient } from "../../common/api/client";
+import { withPaymentsSessionToken } from "../../common/utils/withPaymentsSessionToken";
 import {
   paymentsDeleteMethodAction,
   paymentsGetMethodDetailsAction
 } from "../store/actions";
-import { readablePrivacyReport } from "../../../../utils/reporters";
-import { getGenericError, getNetworkError } from "../../../../utils/errors";
-import { WalletClient } from "../../common/api/client";
-import { walletRemoveCards } from "../../../wallet/store/actions/cards";
-import { mapWalletIdToCardKey } from "../../common/utils";
-import { withPaymentsSessionToken } from "../../common/utils/withPaymentsSessionToken";
 
-/**
- * Handle the remote call to start Wallet onboarding payment methods list
- * @param getPaymentMethods
- * @param action
- */
 export function* handleDeleteWalletDetails(
   deleteWalletById: WalletClient["deleteIOPaymentWalletById"],
   action: ActionType<(typeof paymentsDeleteMethodAction)["request"]>
@@ -33,10 +26,6 @@ export function* handleDeleteWalletDetails(
 
     if (E.isRight(deleteWalletResult)) {
       if (deleteWalletResult.right.status === 204) {
-        yield* put(
-          walletRemoveCards([mapWalletIdToCardKey(action.payload.walletId)])
-        );
-
         // handled success
         const successAction = paymentsDeleteMethodAction.success(
           action.payload.walletId
@@ -66,6 +55,10 @@ export function* handleDeleteWalletDetails(
       action.payload.onFailure?.();
     }
   } catch (e) {
+    const failureAction = paymentsDeleteMethodAction.failure({
+      ...getNetworkError(e)
+    });
+    yield* put(failureAction);
     yield* put(
       paymentsGetMethodDetailsAction.failure({ ...getNetworkError(e) })
     );